src/main/java/rei/java/springboot/controller/EmployeeController.java
package rei.java.springboot.controller;

import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;
import rei.java.springboot.model.Employee;
import rei.java.springboot.service.EmployeeService;

import java.util.List;

// To make it a SpringMvc Controller
@RestController
@RequestMapping("/api/employee")
public class EmployeeController {

    private final EmployeeService employeeService;

    // Constructor based injection
    public EmployeeController(EmployeeService employeeService) {
        this.employeeService = employeeService;
    }

    @PostMapping
    @ResponseStatus(HttpStatus.CREATED)
    public Employee createEmployee(@RequestBody Employee employee) {
        return employeeService.saveEmployee(employee);
    }

    @GetMapping
    public List<Employee> getAllEmployees() {
        return employeeService.getAllEmployee();
    }

    @GetMapping("{id}")
    public ResponseEntity<Employee> getEmployeeById(@PathVariable("id") long id) {
        return employeeService.getEmployeeById(id)
                .map(ResponseEntity::ok)
                .orElseGet(() -> ResponseEntity.notFound().build());
    }

    @PutMapping("{id}")
    public ResponseEntity<Employee> updateEmployee(@PathVariable("id") long id, @RequestBody Employee employee) {
       return employeeService.getEmployeeById(id).map(foundEmployee -> {
           foundEmployee.setFirstName(employee.getFirstName());
           foundEmployee.setLastName(employee.getLastName());
           foundEmployee.setEmail(employee.getEmail());
           Employee updatedEmployee = employeeService.updateEmployee(foundEmployee);

           return new ResponseEntity<>(updatedEmployee, HttpStatus.OK);
       }).orElseGet(() -> ResponseEntity.notFound().build());
    }

    @DeleteMapping("{id}")
    public ResponseEntity<String> deleteEmployee(@PathVariable("id") long employeeId) {
        employeeService.deleteEmployee(employeeId);
        return new ResponseEntity<>("Employee deleted successfully!", HttpStatus.OK);
    }
}
